+++ __init__.py
@@ -1,0 +1,1 @@
+all = ["backend", "database", "schemas", "tests"]
