+++ backend.py
@@ -1,6 +1,8 @@
+import subprocess
+
 from fastapi import FastAPI,Depends
 
-from .database import session_local, engine, User, Article
+from database import session_local, engine, User, Article
 app = FastAPI(title="SUSM Blog", description="A simple blog API", version="0.0.1")
 
 # 未来移入Dependency
@@ -10,6 +12,14 @@ def get_db():
         yield db
     finally:
         db.close()
+
+def launch():
+    """启动后端"""
+    process_backend = subprocess.Popen(
+        'uvicorn backend:app --reload', stdout=subprocess.PIPE, 
+        stderr=subprocess.PIPE,cwd=".",shell=True)
+    print(process_backend.stderr.readline().decode()) # type: ignore
+    return process_backend
 
 @app.get("/")
 async def root():
