+++ database/__init__.py
@@ -3,7 +3,7 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker
 from sqlalchemy import create_engine
 
-SQLALCHEMY_DATABASE_URL = "sqlite:///./backend/test.db"
+SQLALCHEMY_DATABASE_URL = "sqlite:///./database/test.db"
 
 engine = create_engine(
     SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False}
