+++ docs/index.html
@@ -1,1 +1,14 @@
-<h1>测试</h1>
+<!DOCTYPE html>
+<html lang="zh-CN">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="css/index.css" type="text/css">
+    <link rel="shortcut icon" href="img/icon.ico" type="image/x-icon">
+    <title>Test</title>
+</head>
+<body>
+
+</body>
+</html>
